dev: created generateJWT function
