+++ src/main/kotlin/org/openmbee/flexo/mms/auth/plugins/Routing.kt
@@ -1,7 +1,5 @@
 package org.openmbee.flexo.mms.auth.plugins
 
-import com.auth0.jwt.JWT
-import com.auth0.jwt.algorithms.Algorithm
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
 import io.ktor.server.plugins.callloging.*
@@ -24,14 +22,7 @@ fun Application.configureRouting() {
                 val issuer = environment.config.property("jwt.domain").getString()
                 val secret = environment.config.property("jwt.secret").getString()
 
-                val expires = Date(System.currentTimeMillis() + (1 * 24 * 60 * 60 * 1000))
-                val token = JWT.create()
-                    .withAudience(jwtAudience)
-                    .withIssuer(issuer)
-                    .withClaim("username", principal.name)
-                    .withClaim("groups", principal.groups)
-                    .withExpiresAt(expires)
-                    .sign(Algorithm.HMAC256(secret))
+                val token = generateJWT(audience = jwtAudience, issuer = issuer, secret = secret, principal)
                 call.respond(hashMapOf("token" to token))
             }
         }
