+++ src/test/kotlin/org/openmbee/flexo/mms/ApplicationTest.kt
@@ -5,10 +5,10 @@ import io.ktor.client.statement.*
 import io.ktor.http.*
 import io.ktor.server.config.*
 import io.ktor.server.testing.*
-import com.auth0.jwt.algorithms.Algorithm
-import com.auth0.jwt.JWT
 import java.util.*
 import org.junit.jupiter.api.Test
+import org.openmbee.flexo.mms.auth.UserDetailsPrincipal
+import org.openmbee.flexo.mms.auth.plugins.generateJWT
 import org.openmbee.flexo.mms.auth.module
 import kotlin.test.assertEquals
 
@@ -16,7 +16,7 @@ class ApplicationTest {
     private val issuer = "https://localhost/"
     private val audience = "test-audience"
     private val secret = "testsecret"
-    private val relm = "Test Relm"
+    private val realm = "Test Realm"
 
     @Test
     fun userCanAuthenticateWithJWT() = testApplication {
@@ -24,7 +24,7 @@ class ApplicationTest {
         environment {
             config = MapApplicationConfig(
                 "jwt.audience" to audience,
-                "jwt.realm" to relm,
+                "jwt.realm" to realm,
                 "jwt.domain" to issuer,
                 "jwt.secret" to secret,
             )
@@ -34,23 +34,12 @@ class ApplicationTest {
             module()
         }
 
-        val token = generateJWT(username = "test name", groups = listOf("all"))
+        val principal = UserDetailsPrincipal(name = "test name", groups = listOf("all"))
+        val token = generateJWT(issuer = issuer, audience = audience, secret = secret, principal = principal)
         val authTest = client.get("/") {
             header(HttpHeaders.Authorization, "Bearer $token")
         }.bodyAsText()
 
         assertEquals("Hello World!", authTest)
-    }
-
-    private fun generateJWT(username: String, groups: List<String>): String {
-
-        val algorithm = Algorithm.HMAC256(secret)
-        return JWT.create()
-            .withIssuer(issuer)
-            .withAudience(audience)
-            .withClaim("username", username)
-            .withArrayClaim("groups", groups.toTypedArray())
-            .withExpiresAt(Date(System.currentTimeMillis() + 3600000))
-            .sign(algorithm)
     }
 }
